pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "music-teacher-ai"
version = "0.1.0"
description = "Local knowledge base of songs and lyrics to help English teachers find vocabulary in music."
requires-python = ">=3.11"
dependencies = [
    "typer[all]>=0.12",
    "fastapi>=0.111",
    "uvicorn>=0.29",
    "httpx>=0.27",
    "spotipy>=2.23",
    "billboard.py>=2.5",
    "lyricsgenius>=3.0",
    "sqlmodel>=0.0.19",
    "sentence-transformers>=3.0",
    "faiss-cpu>=1.8",
    "nltk>=3.8",
    "apscheduler>=3.10",
    "python-dotenv>=1.0",
    "rich>=13.0",
    "structlog>=24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "black>=24.0",
    "ruff>=0.4",
]

[project.scripts]
music-teacher = "music_teacher_ai.api.cli:main"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "smoke: marks tests that call real external APIs (deselect with -m 'not smoke')",
]
